+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/).
 
 ## Unreleased
+### Fixed
+* Deprecation warnings
 
 ## [3.1](https://pypi.org/project/lupyne/3.1/) - 2023-11-22
 ### Changed